Add files via upload
 This is the first draft of the final project. You will have most of the tools to start the project after Wednesday's lecture on PCA.

You are allowed to use AI to help with coding. You are also allowed to work on the projects with your peers, but you write in your own words.

When explaining your reasoning, be as detailed as possible. Here is a basic tutorial about using Markdown: https://www.datacamp.com/tutorial/markdown-in-jupyter-notebook   Note that you are not graded on aesthetics, but it helps me read if things are organized.

The final upload should be a Juypter notebook

The final section (DNN) may be subject to change, depending on my estimate of the pace of the course. However, I will not change anything after May 4.